+++ .gitignore
@@ -125,3 +125,6 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+
+# levain-graphql-examples specific
+api-cosigner-private-key.json
